requirements.txt
nbtlib==2.0.4
Pillow==10.4.0

